+++ src/tkc/object_array.c
@@ -112,6 +112,7 @@ static ret_t object_array_remove_prop(object_t* obj, const char* name) {
   ret_t ret = RET_NOT_FOUND;
   int32_t index = tk_atoi(name);
   object_array_t* o = OBJECT_ARRAY(obj);
+  return_value_if_fail(isdigit(*name), RET_NOT_FOUND);
 
   if (index >= 0 && index < o->props_size) {
     value_t* iter = o->props + index;
@@ -129,7 +130,7 @@ static ret_t object_array_set_prop(object_t* obj, const char* name, const value_
   int32_t index = tk_atoi(name);
   return_value_if_fail(object_array_extend(obj) == RET_OK, RET_OOM);
 
-  if (index >= 0 && index < o->props_size) {
+  if (isdigit(*name) && index >= 0 && index < o->props_size) {
     value_t* iter = o->props + index;
     value_reset(iter);
     ret = value_deep_copy(iter, v);
@@ -149,7 +150,7 @@ static ret_t object_array_get_prop(object_t* obj, const char* name, value_t* v) 
   if (tk_str_eq(name, "length") || tk_str_eq(name, "size")) {
     value_set_int(v, o->props_size);
     ret = RET_OK;
-  } else {
+  } else if (isdigit(*name)) {
     int32_t index = tk_atoi(name);
     if (index >= 0 && index < o->props_size) {
       value_t* iter = o->props + index;